Make drift-realted strategy
